traefik: fix memory cache tests to run without redis build tag (#234)
Split mesi_test.go so memory cache tests run in the default test suite
without requiring -tags redis. The existing file had //go:build redis which
gated all tests (including memory-only ones) behind the redis build tag.

Changes:
- New cache_memory_test.go: memory cache tests without build tags
- mesi_test.go: redis-only tests with //go:build redis
- Added functional tests: cache initialization, TTL values, config propagation
- Added examples/traefik-memory-cache.yml with memory cache config example
- Added changelog entry for #234
