+++ servers/traefik/mesi_test.go
@@ -10,51 +10,6 @@ import (
 	"time"
 )
 
-func TestCreateConfig(t *testing.T) {
-	config := CreateConfig()
-	if config.MaxDepth != 5 {
-		t.Errorf("Expected MaxDepth 5, got %d", config.MaxDepth)
-	}
-}
-
-func TestNewNilConfig(t *testing.T) {
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})
-	_, err := New(context.Background(), handler, nil, "test")
-	if err == nil {
-		t.Fatal("Expected error for nil config")
-	}
-}
-
-func TestNewDefaultConfig(t *testing.T) {
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})
-	config := CreateConfig()
-	p, err := New(context.Background(), handler, config, "test")
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-	if p == nil {
-		t.Fatal("Expected non-nil plugin")
-	}
-}
-
-func TestNewMemoryCache(t *testing.T) {
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})
-	config := CreateConfig()
-	config.CacheBackend = "memory"
-	config.CacheTTL = "60s"
-	config.CacheSize = 100
-
-	p, err := New(context.Background(), handler, config, "test")
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-
-	plugin := p.(*ResponsePlugin)
-	if plugin.cache == nil {
-		t.Fatal("Expected non-nil cache")
-	}
-}
-
 func TestNewRedisCache(t *testing.T) {
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})
 	config := CreateConfig()
@@ -110,44 +65,7 @@ func TestNewRedisCacheWithPassword(t *testing.T) {
 	}
 }
 
-func TestNewInvalidCacheBackend(t *testing.T) {
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})
-	config := CreateConfig()
-	config.CacheBackend = "invalid"
-
-	_, err := New(context.Background(), handler, config, "test")
-	if err == nil {
-		t.Fatal("Expected error for invalid cache backend")
-	}
-}
-
-func TestNewInvalidCacheTTL(t *testing.T) {
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})
-	config := CreateConfig()
-	config.CacheBackend = "memory"
-	config.CacheTTL = "invalid"
-
-	_, err := New(context.Background(), handler, config, "test")
-	if err == nil {
-		t.Fatal("Expected error for invalid cache TTL")
-	}
-}
-
-func TestName(t *testing.T) {
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})
-	config := CreateConfig()
-	p, err := New(context.Background(), handler, config, "test")
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-
-	plugin := p.(*ResponsePlugin)
-	if plugin.Name() != "mesi" {
-		t.Errorf("Expected name 'mesi', got %s", plugin.Name())
-	}
-}
-
-func TestClose(t *testing.T) {
+func TestCloseRedis(t *testing.T) {
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})
 	config := CreateConfig()
 	config.CacheBackend = "redis"
@@ -166,82 +84,31 @@ func TestClose(t *testing.T) {
 	}
 }
 
-func TestServeHTTPWithCache(t *testing.T) {
+func TestRedisCacheWithConfig(t *testing.T) {
 	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "text/html")
 		w.WriteHeader(http.StatusOK)
-		w.Write([]byte("<html><body><esi:include src=\"/fragment\" /></body></html>"))
+		w.Write([]byte("<html><body>content</body></html>"))
 	})
 
 	config := CreateConfig()
-	config.CacheBackend = "memory"
-	config.CacheTTL = "60s"
+	config.CacheBackend = "redis"
+	config.CacheTTL = "120s"
+	config.CacheRedisAddr = "10.0.0.5:6379"
+	config.CacheRedisPassword = "password"
+	config.CacheRedisDB = 2
 
 	p, err := New(context.Background(), handler, config, "test")
 	if err != nil {
 		t.Fatalf("Unexpected error: %v", err)
 	}
 
-	req := httptest.NewRequest("GET", "http://example.com/", nil)
-	rec := httptest.NewRecorder()
-
-	p.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusOK {
-		t.Errorf("Expected status 200, got %d", rec.Code)
+	plugin := p.(*ResponsePlugin)
+	if plugin.cacheTTL != 120*time.Second {
+		t.Errorf("Expected cacheTTL 120s, got %v", plugin.cacheTTL)
 	}
-}
-
-func TestServeHTTPWithoutCache(t *testing.T) {
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "text/html")
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte("<html><body><esi:include src=\"/fragment\" /></body></html>"))
-	})
-
-	config := CreateConfig()
-
-	p, err := New(context.Background(), handler, config, "test")
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-
-	req := httptest.NewRequest("GET", "http://example.com/", nil)
-	rec := httptest.NewRecorder()
-
-	p.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusOK {
-		t.Errorf("Expected status 200, got %d", rec.Code)
-	}
-}
-
-func TestServeHTTPNonHTML(t *testing.T) {
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte(`{"status":"ok"}`))
-	})
-
-	config := CreateConfig()
-	config.CacheBackend = "memory"
-	config.CacheTTL = "60s"
-
-	p, err := New(context.Background(), handler, config, "test")
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-
-	req := httptest.NewRequest("GET", "http://example.com/api", nil)
-	rec := httptest.NewRecorder()
-
-	p.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusOK {
-		t.Errorf("Expected status 200, got %d", rec.Code)
-	}
-	if rec.Body.String() != `{"status":"ok"}` {
-		t.Errorf("Expected body {'status':'ok'}, got %s", rec.Body.String())
+	if plugin.cache == nil {
+		t.Fatal("Expected non-nil cache")
 	}
 }
 
@@ -274,33 +141,5 @@ func TestCacheIntegration(t *testing.T) {
 
 	if rec1.Body.String() != rec2.Body.String() {
 		t.Errorf("Expected same body for cached response")
-	}
-}
-
-func TestRedisCacheWithConfig(t *testing.T) {
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "text/html")
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte("<html><body>content</body></html>"))
-	})
-
-	config := CreateConfig()
-	config.CacheBackend = "redis"
-	config.CacheTTL = "120s"
-	config.CacheRedisAddr = "10.0.0.5:6379"
-	config.CacheRedisPassword = "password"
-	config.CacheRedisDB = 2
-
-	p, err := New(context.Background(), handler, config, "test")
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-
-	plugin := p.(*ResponsePlugin)
-	if plugin.cacheTTL != 120*time.Second {
-		t.Errorf("Expected cacheTTL 120s, got %v", plugin.cacheTTL)
-	}
-	if plugin.cache == nil {
-		t.Fatal("Expected non-nil cache")
 	}
 }
